+++ resources/views/back/_layouts/_editors/redactor/editor.blade.php
@@ -26,7 +26,9 @@
                     }
                 }
             },
-            imageUpload: '{{ $imageUploadUrl }}',
+            @if(isset($imageUploadUrl) && (!isset($disableImageUpload) || !$disableImageUpload))
+                imageUpload: '{{ $imageUploadUrl }}',
+            @endif
             definedlinks: '{{ route('chief.api.internal-links') }}',
             customClasses: [
                 {