+++ src/HealthMonitor/Checks/HomepageSetCheck.php
@@ -5,6 +5,7 @@ declare(strict_types = 1);
 namespace Thinktomorrow\Chief\HealthMonitor\Checks;
 
 use Thinktomorrow\Chief\Settings\Setting;
+use Thinktomorrow\Chief\HealthMonitor\Notifiers\AlertBarNotifier;
 
 class HomepageSetCheck implements HealthCheck
 {
@@ -18,5 +19,12 @@ class HomepageSetCheck implements HealthCheck
     public function message(): string
     {
         return 'Het lijkt erop dat er geen homepagina ingesteld is. Stel er een in hier: <a href="'. route('chief.back.settings.edit') .'" class="text-secondary-800 underline hover:text-white">Settings</a>';
+    }
+
+    public function notifiers(): array
+    {
+        return [
+            AlertBarNotifier::class,
+        ];
     }
 }