+++ src/HealthMonitor/Exceptions/InvalidClassException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Thinktomorrow\Chief\HealthMonitor\Exceptions;
+
+class InvalidClassException extends \Exception
+{
+}