+++ src/Fields/Types/FieldType.php
@@ -16,7 +16,8 @@ class FieldType
     const SELECT = 'select';  // Select options
     const MEDIA = 'media';  // media file (slim uploader)
     const DOCUMENT = 'document';  // documents
-    const RADIO = 'radio';  // media file (slim uploader)
+    const RADIO = 'radio';  // radio select
+    const CHECKBOX = 'checkbox';  // checkbox select
     const PAGEBUILDER = 'pagebuilder';  // the most special field there is...
 
     /** @var string */